Apply all review feedback: JSON escaping, input validation, wh_read bug fix, exit_code consistency
Agent-Logs-Url: https://github.com/sjpupro-lab/Canvas-os/sessions/9a0bef88-8615-4c18-9567-37758a483eb0

+++ engine/include/canvas_gpu.h
@@ -35,7 +35,7 @@ int  gpu_scan_active_set(GpuCtx* g, uint32_t* out, int cap);
 typedef struct {
     uint32_t region_start;
     uint32_t region_len;
-    uint32_t hash;        /* FNV-1a of region */
+    uint64_t hash;        /* FNV-1a of region (64-bit, DK-5) */
 } BhSummary;
 
 int  gpu_bh_summarize_idle(GpuCtx* g, BhSummary* out, int cap);
